requirements.txt
streamlit>=1.28.0
snowflake-snowpark-python>=1.0.0
pandas>=1.5.0
